implement deck reorder function with drag and drop interface

+++ src/App/Deck.purs
@@ -5,21 +5,29 @@ import Prelude
 import Control.Monad.Maybe.Trans (runMaybeT)
 import Data.Array ((!!))
 import Data.Array as Array
-import Data.Foldable (fold)
+import Data.Foldable (fold, traverse_)
 import Data.Function (on)
 import Data.Int as Int
 import Data.Maybe (fromMaybe)
+import Data.MediaType (MediaType(..))
 import Data.Monoid.Additive (Additive(..))
 import Data.Newtype (alaF, collect)
 import Data.String as String
 import Effect.Aff (Aff)
+import Halogen (RefLabel(..))
 import Halogen as H
 import Halogen.HTML as HH
 import Halogen.HTML.Events as HE
 import Halogen.HTML.Properties as HP
-import TcgCalculator.Types (Card, Deck, generateId)
+import TcgCalculator.Types (Card, Deck, Id, generateId)
+import TcgCalculator.Types.Id as Id
 import Util.Array as ArrayUtil
 import Util.Halogen as HU
+import Web.Event.Event as Event
+import Web.HTML (window)
+import Web.HTML.Event.DataTransfer as DataTransfer
+import Web.HTML.Event.DragEvent as Drag
+import Web.HTML.Window (confirm)
 
 ----------------------------------------------------------------
 
@@ -32,7 +40,9 @@ data Action
   | UpdateDeck Int
   | UpdateHand Int
   | UpdateOthers Int
-  | Swap Index Index
+  | StartReorder Id Drag.DragEvent
+  | HandleDragBehavior Drag.DragEvent
+  | ExecuteReorder Id Drag.DragEvent
 
 data Query a
   = SetDeck Deck a
@@ -56,7 +66,7 @@ component = H.mkComponent
     HH.div
       [ HP.class_ $ H.ClassName "p-1 rounded border-2 border-amber-500" ]
       [ renderHeader deckCount hand cardCount
-      , HH.ul [ HP.class_ $ H.ClassName "m-1" ] $ renderCard others <$> cards
+      , renderCardList others cards
       , renderFooter others cardCount
       ]
 
@@ -94,26 +104,47 @@ component = H.mkComponent
           ]
       ]
 
+  renderCardList others cards =
+    HH.ul
+      [ HP.class_ $ H.ClassName "m-1"
+      , HE.onDragEnter HandleDragBehavior
+      , HE.onDragOver HandleDragBehavior
+      ]
+      $ renderCard others <$> cards
+
   renderCard others card =
     HH.li
-      [ HP.class_ $ H.ClassName "flex border-b border-gray-500" ]
+      [ HP.class_ $ H.ClassName "flex"
+      , HP.ref $ H.RefLabel (Id.toString card.id)
+      , HE.onDrop $ ExecuteReorder card.id
+      ]
       [ HH.div
-          [ HP.class_ $ H.ClassName "mx-1" ]
-          [ HU.removeButton (RemoveCard card) ]
-      , HH.input
-          [ HP.classes [ H.ClassName "grow", styleFormInput ]
-          , HP.type_ HP.InputText
-          , HP.value card.name
-          , HE.onValueChange (UpdateCard <<< card { name = _ })
+          [ HP.class_ $ H.ClassName "flex items-center mx-1" ]
+          [ HH.div
+              [ HP.class_ $ H.ClassName "hidden md:block px-1 cursor-grab active:cursor-grabbing text-gray-500"
+              , HP.draggable true
+              , HE.onDragStart $ StartReorder card.id
+              ]
+              [ HU.fa_ "fa-grip-vertical" ]
+          , HU.removeButton (RemoveCard card)
           ]
-      , HH.input
-          [ HP.class_ styleFormNumber
-          , HP.type_ HP.InputNumber
-          , HP.step $ HP.Step 1.0
-          , HP.value $ show card.count
-          , HP.min 0.0
-          , HP.max if String.null card.name then 0.0 else Int.toNumber (card.count + others)
-          , HE.onValueChange (UpdateCard <<< card { count = _ } <<< fromMaybe 0 <<< Int.fromString)
+      , HH.div
+          [ HP.class_ $ H.ClassName "grow flex border-b border-gray-500" ]
+          [ HH.input
+              [ HP.classes [ H.ClassName "grow", styleFormInput ]
+              , HP.type_ HP.InputText
+              , HP.value card.name
+              , HE.onValueChange (UpdateCard <<< card { name = _ })
+              ]
+          , HH.input
+              [ HP.class_ styleFormNumber
+              , HP.type_ HP.InputNumber
+              , HP.step $ HP.Step 1.0
+              , HP.value $ show card.count
+              , HP.min 0.0
+              , HP.max if String.null card.name then 0.0 else Int.toNumber (card.count + others)
+              , HE.onValueChange (UpdateCard <<< card { count = _ } <<< fromMaybe 0 <<< Int.fromString)
+              ]
           ]
       ]
 
@@ -154,10 +185,11 @@ component = H.mkComponent
       H.modify_ do
         cards <- _.cards
         _ { cards = Array.snoc cards { id, name: "", count: 0 } }
-    RemoveCard card -> do
-      { cards, others } <- H.get
-      let cards' = Array.deleteBy ((==) `on` _.id) card cards
-      raiseUpdate =<< H.modify _ { cards = cards', others = others + card.count }
+    RemoveCard card -> void $ H.fork do
+      whenM (if String.null card.name then pure true else H.liftEffect $ confirm ("カード「" <> card.name <> "」を削除します。") =<< window) do
+        { cards, others } <- H.get
+        let cards' = Array.deleteBy ((==) `on` _.id) card cards
+        raiseUpdated =<< H.modify _ { cards = cards', others = others + card.count }
     UpdateCard card -> do
       { cards, others } <- H.get
       fold do
@@ -165,28 +197,49 @@ component = H.mkComponent
         old <- cards !! i
         let new = if String.null card.name then card { count = 0 } else card { count = clamp 0 (old.count + others) card.count }
         cards' <- Array.updateAt i new cards
-        pure $ raiseUpdate =<< H.modify _ { cards = cards', others = others - (new.count - old.count) }
+        pure $ raiseUpdated =<< H.modify _ { cards = cards', others = others - (new.count - old.count) }
     UpdateDeck total -> do
       cards <- H.gets _.cards
       let cardCount = countCards cards
       action $ UpdateOthers (total - cardCount)
     UpdateHand hand -> do
-      raiseUpdate =<< H.modify do
+      raiseUpdated =<< H.modify do
         { cards, others } <- identity
         let deckCount = countCards cards + others
         _ { hand = clamp 1 deckCount hand }
     UpdateOthers others -> do
-      raiseUpdate =<< H.modify do
+      raiseUpdated =<< H.modify do
         { cards, hand } <- identity
         let cardCount = countCards cards
         let deckCount = clamp cardCount deckLimit (cardCount + others)
         _ { others = deckCount - cardCount, hand = min hand deckCount }
-    Swap x y -> do -- TODO
-      H.modify_ do
-        cards <- _.cards
-        _ { cards = ArrayUtil.swap x y cards }
+    StartReorder id event -> do
+      let transfer = Drag.dataTransfer event
+      H.liftEffect $ DataTransfer.setData (MediaType dragItemMediaType) (Id.toString id) transfer
+      elem <- H.getRef $ RefLabel (Id.toString id)
+      elem # traverse_ \e -> do
+        H.liftEffect $ DataTransfer.setDragImage transfer e 5 15
+    HandleDragBehavior event -> do
+      let transfer = Drag.dataTransfer event
+      when (DataTransfer.types transfer == [dragItemMediaType]) do
+        H.liftEffect <<< Event.preventDefault $ Drag.toEvent event
+    ExecuteReorder destination event -> do
+      let transfer = Drag.dataTransfer event
+      id <- H.liftEffect $ DataTransfer.getData (MediaType dragItemMediaType) transfer
+      unless (String.null id) do
+        raiseUpdated =<< H.modify do
+          cards <- _.cards
+          let cards' = fold do
+                target <- Id.fromString id
+                from <- Array.findIndex (_.id >>> (_ == target)) cards
+                to <- Array.findIndex (_.id >>> (_ == destination)) cards
+                pure $ ArrayUtil.shiftInsert from to cards
+          _ { cards = cards' }
+        H.liftEffect <<< Event.preventDefault $ Drag.toEvent event
 
-  raiseUpdate deck =
+  dragItemMediaType = "tcg-calculator/card"
+
+  raiseUpdated deck =
     H.raise deck { cards = Array.filter (_.name >>> not String.null) deck.cards }
 
   countCards = alaF Additive Array.foldMap _.count
